@@ -4,10 +4,12 @@
   "private": true,
   "dependencies": {
     "@apollo/client": "^3.4.9",
-    "@craco/craco": "^6.2.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@craco/craco": "^6.4.3",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/lab": "^5.0.0-alpha.63",
+    "@mui/material": "^5.2.7",
+    "@mui/styled-engine-sc": "^5.1.0",
+    "@mui/styles": "^5.2.3",
     "@svgdotjs/svg.draggable.js": "^3.0.2",
     "@svgdotjs/svg.js": "^3.1.1",
     "@testing-library/jest-dom": "^5.14.1",
@@ -17,21 +19,22 @@
     "formik": "^2.2.9",
     "graphql": "^15.5.1",
     "history": "^4.10.1",
-    "html-to-image": "^1.7.0",
+    "html-to-image": "^1.9.0",
+    "material-ui-color": "^1.2.0",
     "react": "^17.0.2",
     "react-cool-dimensions": "^2.0.7",
     "react-copy-to-clipboard": "^5.0.4",
-    "react-dnd": "^14.0.3",
-    "react-dnd-html5-backend": "^14.0.1",
+    "react-dnd": "^14.0.5",
+    "react-dnd-html5-backend": "^14.1.0",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.1",
+    "react-scripts": "^5.0.0",
+    "styled-components": "^5.3.3",
     "subscriptions-transport-ws": "^0.9.19",
-    "twin.macro": "^2.7.0",
-    "typescript": "^4.4.2",
-    "web-vitals": "^2.1.0",
-    "yup": "^0.32.9"
+    "twin.macro": "^2.8.2",
+    "typescript": "^4.5.4",
+    "web-vitals": "^2.1.3",
+    "yup": "^0.32.11"
   },
   "scripts": {
     "start": "craco start",
@@ -73,26 +76,25 @@
     }
   },
   "devDependencies": {
-    "@tailwindcss/postcss7-compat": "^2.2.8",
     "@types/history": "^4.7.9",
     "@types/jest": "^26.0.24",
     "@types/node": "^15.12.2",
     "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
     "@types/react-copy-to-clipboard": "^5.0.0",
+    "@types/react-dom": "^17.0.9",
     "@types/react-router-dom": "^5.1.8",
-    "@typescript-eslint/eslint-plugin": "^4.29.3",
-    "@typescript-eslint/parser": "^4.29.3",
+    "@typescript-eslint/eslint-plugin": "^5.9.0",
+    "@typescript-eslint/parser": "^5.9.0",
     "apollo": "^2.33.4",
-    "autoprefixer": "^9.8.6",
+    "autoprefixer": "^10.4.2",
     "craco-alias": "^3.0.1",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "postcss": "^7.0.36",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.1.4"
+    "eslint": "^8.6.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "postcss": "^8.4.5",
+    "tailwindcss": "^3.0.12"
   }
 }
