@@ -1,11 +1,11 @@
 import React from 'react';
 
-import { CircularProgress } from '@material-ui/core';
+import { CircularProgress } from '@mui/material';
 
 export default function LoadingIndicator() {
   return (
     <div className="flex justify-center my-4">
-      <CircularProgress color="secondary" />
+      <CircularProgress color="primary" />
     </div>
   );
 }
