@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 
-import { Button } from '@material-ui/core';
-import GroupAddIcon from '@material-ui/icons/GroupAdd';
+import GroupAddIcon from '@mui/icons-material/GroupAdd';
+import { Button } from '@mui/material';
 import { CopyToClipboard } from 'react-copy-to-clipboard';
 
 const BUTTON_TEXTS = {
@@ -30,10 +30,10 @@ export default function TeamCodeCopyButton({ teamCode }: { teamCode: string }) {
       <Button
         variant="outlined"
         className="mb-2 w-full"
-        color="secondary"
+        color="primary"
         onMouseEnter={() => !copied && setButtonText(BUTTON_TEXTS.hover)}
         onMouseLeave={() => !copied && setButtonText(BUTTON_TEXTS.normal)}
-        startIcon={!copied && <GroupAddIcon color="secondary" fontSize="large" />}
+        startIcon={!copied && <GroupAddIcon color="primary" fontSize="large" />}
       >
         {buttonText}
       </Button>
