@@ -1,8 +1,9 @@
 import React, {
+  memo,
   MutableRefObject, useEffect, useRef, useState,
 } from 'react';
 
-import { Button } from '@material-ui/core';
+import { Button } from '@mui/material';
 import {
   SVG, Svg, Path, Circle as CircleBase, G, extend,
 } from '@svgdotjs/svg.js';
@@ -24,6 +25,7 @@ type Props = {
   printMode?: boolean;
   onSave?: (updatedItems: UpdatedItemsMap) => void;
   onCancel?: () => void;
+  labelClassName?: string;
 }
 
 type HillSvg = {
@@ -52,15 +54,17 @@ export const VIEW_BOX = DEFAULT_VIEW_BOX;
 
 const DOT_DIAMETER = 10;
 const DOT_RADIUS = DOT_DIAMETER / 2;
+let createSvgTimeout: number;
 
-export default function HillChart({
-  width, height, allowEdit, printMode, onSave, onCancel, data = [],
+function HillChartComponent({
+  width, height, allowEdit, printMode, onSave, onCancel, data = [], labelClassName,
 }: Props) {
   const chartId = useRef<number>(Math.floor(Math.random() * Math.floor(101)));
   const updatedItems = useRef<UpdatedItemsMap>({});
   const hillChartSvg = useRef<HillSvg>();
   const container = useRef<HTMLDivElement>(null);
   const [plottedItems, setPlottedItems] = useState<string[]>([]);
+  const [svgReady, setSvgReady] = useState(false);
 
   const plotPoints = () => {
     const circles = currentChartPoints(hillChartSvg.current);
@@ -112,7 +116,13 @@ export default function HillChart({
   useEffect(() => {
     if (container.current) {
       hillChartSvg.current = createSvg(container.current);
+      // This kinda sucks...
+      // The SVG needs to be rendered in the DOM before any labels are added
+      // to avoid flicker.
+      createSvgTimeout = window.setTimeout(() => setSvgReady(true), 500);
     }
+    // eslint-disable-next-line no-unused-expressions
+    return () => { createSvgTimeout && clearTimeout(createSvgTimeout); };
   }, [container]);
 
   useEffect(() => {
@@ -124,7 +134,7 @@ export default function HillChart({
     const updateAllLabelPos = () => plottedItems.forEach((itemId) => {
       const point = findChartPoint(currentChartPoints(hillChartSvg.current), itemId);
       if (point) {
-        updatePointLabelPos(point, VIEW_BOX);
+        updatePointLabelPos(hillChartSvg.current!.canvas, point, VIEW_BOX);
       }
     });
 
@@ -133,9 +143,10 @@ export default function HillChart({
   }, [plottedItems]);
 
   useEffect(() => {
+    if (!hillChartSvg.current) { return; }
     const points = currentChartPoints(hillChartSvg.current);
     if (allowEdit) {
-      points.forEach((point) => enableDragForPoint(point, updatedItems));
+      points.forEach((point) => enableDragForPoint(hillChartSvg.current!, point, updatedItems));
     } else {
       points.forEach((point) => disableDragForPoint(point));
       updatedItems.current = {};
@@ -144,19 +155,31 @@ export default function HillChart({
 
   return (
     <div ref={container} id={`chart-container-${chartId.current}`} className="relative" style={{ width, height }}>
-      {plottedItems.map((itemId) => {
+      {svgReady && plottedItems.map((itemId) => {
         const point = findChartPoint(currentChartPoints(hillChartSvg.current), itemId);
-        return point && <ChartPointLabel key={itemId} point={point} dragEnabled={allowEdit} printMode={printMode} />;
+        return point && (
+          <ChartPointLabel
+            key={itemId}
+            chart={hillChartSvg.current!.canvas}
+            point={point}
+            dragEnabled={allowEdit}
+            printMode={printMode}
+            className={labelClassName}
+          />
+        );
       })}
       {allowEdit && (
         <div className="absolute top-2 right-2 flex">
-          <Button className="mr-2" variant="outlined" onClick={handleCancel}>Cancel</Button>
-          <Button className="text-white" variant="contained" color="secondary" onClick={handleSave}>Save</Button>
+          <Button className="mr-2" variant="outlined" color="secondary" onClick={handleCancel}>Cancel</Button>
+          <Button className="text-white" variant="contained" color="primary" onClick={handleSave}>Save</Button>
         </div>
       )}
     </div>
   );
 }
+
+const HillChart = memo(HillChartComponent);
+export default HillChart;
 
 function currentChartPoints(svg?: HillSvg) {
   return (svg?.pointsGroup.children().toArray() || []) as CircleElement[];
@@ -225,14 +248,14 @@ function addChartPoint(
     .draggable();
 
   if (enableDrag) {
-    enableDragForPoint(point, updatedItemsRef);
+    enableDragForPoint(svg, point, updatedItemsRef);
   } else {
     point.on('beforedrag.disabled', (event: any) => {
       event.preventDefault();
     });
   }
 
-  updatePointLabelPos(point, VIEW_BOX);
+  updatePointLabelPos(svg.canvas, point, VIEW_BOX);
   return point;
 }
 
@@ -245,7 +268,7 @@ function disableDragForPoint(point: Circle | CircleElement) {
   });
 }
 
-function enableDragForPoint(point: Circle | CircleElement, updatedItemsRef: MutableRefObject<UpdatedItemsMap>) {
+function enableDragForPoint(svg: HillSvg, point: Circle | CircleElement, updatedItemsRef: MutableRefObject<UpdatedItemsMap>) {
   point.off('beforedrag.disabled');
   point.off('dragmove.progressUpdate');
 
@@ -263,6 +286,6 @@ function enableDragForPoint(point: Circle | CircleElement, updatedItemsRef: Muta
     // eslint-disable-next-line no-param-reassign
     updatedItemsRef.current[point.chartItem.id] = getProgressFromPosition(box.cx); // Bleh, stop mutating this arg
     handler.move(moveX - DOT_RADIUS, VIEW_BOX.y - hillForumula(moveX, VIEW_BOX) - DOT_RADIUS);
-    updatePointLabelPos(point, VIEW_BOX);
+    updatePointLabelPos(svg.canvas, point, VIEW_BOX);
   });
 }
