@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 
-import { Typography } from '@material-ui/core';
+import { Typography } from '@mui/material';
 
 import useUpdateProject from '@/api/mutations/useUpdateProject';
 import { Projects_projects as Project } from '@/api/queries/types/Projects';
