@@ -1,5 +1,6 @@
 import React, { useEffect, useRef, useState } from 'react';
 
+import { Svg } from '@svgdotjs/svg.js';
 import tw, { styled, css } from 'twin.macro';
 
 import {
@@ -24,13 +25,22 @@ const StyledLabel = styled.div`
   }
 `;
 
-type Props = { point: CircleElement, dragEnabled?: boolean, printMode?: boolean };
+type Props = {
+  chart: Svg;
+  point: CircleElement;
+  dragEnabled?: boolean;
+  printMode?: boolean;
+  className?: string;
+};
 
 type EventListener = (event: Event) => void;
 type Position = { top: number, left: number, bottom: number, right: number, progress: number };
-export default function ChartPointLabel({ point, dragEnabled, printMode }: Props) {
+export default function ChartPointLabel({
+  chart, point, dragEnabled, printMode, className = '',
+}: Props) {
+  const chartRect = chart.node.getBoundingClientRect();
   const startPos = point.node.getBoundingClientRect();
-  const [pos, setPos] = useState<Position>(getStartPosition(startPos, point));
+  const [pos, setPos] = useState<Position>(getStartPosition(chartRect, startPos, point));
   const [labelStyle, setLabelStyle] = useState<React.CSSProperties>({
     top: startPos.top - 5,
     left: startPos.left + 20,
@@ -40,7 +50,7 @@ export default function ChartPointLabel({ point, dragEnabled, printMode }: Props
   // Bleh, the start position isn't always initialized to real values at first.
   // TODO: Investigate this more...
   useEffect(() => {
-    setPos(getStartPosition(startPos, point));
+    setPos(getStartPosition(chartRect, startPos, point));
   }, [startPos.top, startPos.bottom, startPos.left, startPos.right]);
 
   useEffect(() => {
@@ -85,7 +95,7 @@ export default function ChartPointLabel({ point, dragEnabled, printMode }: Props
     <StyledLabel
       ref={labelEl}
       id={`${point.chart}.item.${point.chartItem.id}.label`}
-      className={`fixed ${dragEnabled ? 'cursor-move' : ''} ${inProgress ? 'text-gray-800' : 'text-gray-300'}`}
+      className={`${className} absolute ${dragEnabled ? 'cursor-move' : ''} ${inProgress ? 'text-gray-800' : 'text-gray-300'}`}
       style={labelStyle}
       $printMode={printMode}
     >
@@ -98,32 +108,31 @@ function chartPositionForPoint(point: Circle | CircleElement) {
   return parseInt(point.node.getAttribute('cx') || '0', 10);
 }
 
-export function updatePointLabelPos(point: Circle | CircleElement, viewbox: ViewBox) {
-  const {
-    top, left, bottom, right,
-  } = point.node.getBoundingClientRect();
+export function updatePointLabelPos(chart: Svg, point: Circle | CircleElement, viewbox: ViewBox) {
+  const pointRect = point.node.getBoundingClientRect();
+  const chartRect = chart.node.getBoundingClientRect();
   const chartPos = chartPositionForPoint(point);
   const moveEvent = new CustomEvent(`${point.chart}.item.${point.chartItem.id}`, {
     detail: {
-      top,
-      left,
-      bottom,
-      right,
+      top: pointRect.top - chartRect.top,
+      left: pointRect.left - chartRect.left,
+      bottom: pointRect.bottom - chartRect.top,
+      right: pointRect.right - chartRect.left,
       progress: getProgressFromPosition(chartPos, viewbox),
     },
   });
   document.getElementById(`${point.chart}.item.${point.chartItem.id}.label`)?.dispatchEvent(moveEvent);
 }
 
-function getStartPosition(startPos: DOMRect, point: CircleElement) {
+function getStartPosition(chart: DOMRect, startPos: DOMRect, point: CircleElement) {
   // document.elementsFromPoint(x, y).map((e) => e.id).filter(/* *.item.*.label */)
   // ^ check for 3 points: top edge, bottom edge, middle
   // if any elements already exist, dodge somehow
   return {
-    top: startPos.top,
-    left: startPos.left,
-    bottom: startPos.bottom,
-    right: startPos.right,
+    top: startPos.top - chart.top,
+    left: startPos.left - chart.left,
+    bottom: startPos.bottom - chart.top,
+    right: startPos.right - chart.left,
     progress: getProgressFromPosition(chartPositionForPoint(point)),
   };
 }
