@@ -8,8 +8,8 @@ for (const [key, value] of breakpointEntries) {
 }
 
 module.exports = {
-  purge: ['./src/**/*.{js,jsx,ts,tsx}'],
-  darkMode: false, // or 'media' or 'class'
+  content: ['./src/**/*.{js,jsx,ts,tsx}'],
+  darkMode: 'media', // or 'media' or 'class'
   theme: {
     fontFamily: {
       poppins: ['Poppins'],
